x86: export x86 boot_params to sysfs
kexec-tools use boot_params for getting the 1st kernel hardware_subarch,
the kexec kernel efi runtime support also need read the old efi_info from
boot_params. Currently it exists in debugfs which is not a good place for
such infomation. Per HPA, we should avoid of "sploit debugfs".

In this patch /sys/kernel/boot_params are exported, also the setup_data
is exported as a subdirectory. For original debugfs since it's already
there for long time and kexec-tools is using it for hardware_subarch so
let's do not remove them for now.

Structure are like below:

/sys/kernel/boot_params
|__ data                /* binary data for boot_params */
|__ setup_data          /* subdirectory for setup_data if there's any */
|   |__ 0               /* the first setup_data node */
|   |   |__ data        /* binary data for setup_data node 0 */
|   |   |__ type        /* setup_data type of setup_data node 0, hex string */
[snip]
|__ version             /* hex string for boot protocal version */

Changelog:
Greg: use __ATTR_RO() and group attr.

Signed-off-by: Dave Young <dyoung@redhat.com>